requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
requests>=2.31.0
wikipedia>=1.4.0
rich>=13.7.0
python-dotenv>=1.0.0
